fix type error
> PHP Fatal error:  Uncaught TypeError: abs(): Argument #1 ($number) must
> be of type int|float, string given in Yaml/Parser.php:463
>
> #0 Yaml/Parser.php(463): abs('')  \
> #1 ...

Fix is to move the int cast of the return value of the abs() function
to the parameter of it.

yaml spec is different here, it does "- 0" for the integer calculation
and no integer conversion when empty (the case here), this parse behavior
is kept unchanged.

in yaml this is "8.1.1.1. Block Indentation Indicator"

    [163] 	c-indentation-indicator(m) 	::=
       ns-dec-digit ⇒ m = ns-dec-digit - #x30
       /* Empty */  ⇒ m = auto-detect()


PHP: 8.0 beta 3
Log: https://travis-ci.org/github/ktomk/pipelines/jobs/727223470#L1898
Yaml-Spec: https://yaml.org/spec/1.2/spec.html#id2793979